fix hist stats error
as spec said, every hist bin is more than 16 bits, so change the type from
unsigned short to unsigned int.
notice that this commit should be applied with the corresponded kernel commit.

+++ rkisp/ia-engine/include/linux/media/rkisp1-config_v12.h
@@ -728,7 +728,7 @@ struct cifisp_af_stat {
  * with ISP_HIST_XXX
  */
 struct cifisp_hist_stat {
-	unsigned short hist_bins[CIFISP_HIST_BIN_N_MAX];
+	unsigned int hist_bins[CIFISP_HIST_BIN_N_MAX];
 } __attribute__ ((packed));
 
 /**
